requirements.txt
python-dotenv==1.1.1
requests==2.32.4
mkdocs-material==9.6.15
mkdocstrings-python==1.16.12
duckdb==1.3.2